pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-anywhere"
version = "0.1.0"
description = "Secure, portable, local-first RAG system with configurable embedding models"
authors = [
    {name = "Andrew", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["rag", "embedding", "vector-search", "document-search", "nlp"]

dependencies = [
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]
openai = [
    "openai>=1.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all = [
    "rag-anywhere[api,openai,dev]",
]

[project.scripts]
rag-anywhere = "rag_anywhere.cli.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/rag-anywhere"
Repository = "https://github.com/yourusername/rag-anywhere"
Issues = "https://github.com/yourusername/rag-anywhere/issues"

[tool.setuptools]
packages = ["rag_anywhere"]

[tool.setuptools.package-data]
rag_anywhere = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
